Add MUI, change Directories
