+++ package.json
@@ -3,9 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/material": "^5.12.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "install": "^0.13.0",
+    "npm": "^9.6.4",
     "react": "^18.2.0",
     "react-cookie": "^4.1.1",
     "react-cookies": "^0.1.1",
