+++ src/components/Main.js
@@ -1,13 +1,13 @@
-import LogIn from "./LogIn";
+import LogIn from "./partials/LogIn";
 import MainContents from "./MainContents";
-import {useContext} from "react";
+import { useContext } from "react";
 
-const Main = ({isLoggedIn, handleLogIn}) => {
-    if (isLoggedIn) {
-        return <MainContents />;
-    } else {
-        return <LogIn handleLogIn={handleLogIn} />;
-    }
+const Main = ({ isLoggedIn, handleLogIn }) => {
+  if (isLoggedIn) {
+    return <MainContents />;
+  } else {
+    return <LogIn handleLogIn={handleLogIn} />;
+  }
 };
 
 export default Main;
