+++ src/components/partials/LogIn.js
@@ -1,0 +1,89 @@
+import { Link } from "react-router-dom";
+import { useState } from "react";
+import { API } from "../../utils/config";
+import * as status from "../../utils/status";
+import CircularBackdrop from "../ui/CircularBackdrop";
+
+const LogIn = ({ handleLogIn }) => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [open, setOpen] = useState(false);
+
+  const handleOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  const handleIdInput = (e) => {
+    setUsername(e.target.value);
+  };
+
+  const handlePwInput = (e) => {
+    setPassword(e.target.value);
+  };
+
+  const getLogInResult = async (data) => {
+    const response = await fetch(`${API.LOGIN}`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      credentials: "include",
+      body: JSON.stringify(data),
+    });
+    // respones의 status를 확인해서 상황에 알맞은 Error를 던집니다.
+    status.loginStatus(response.status);
+    return response.json();
+  };
+
+  const handleSubmit = async (error) => {
+    error.preventDefault();
+    handleOpen();
+    const data = { userid: username, passwd: password };
+    try {
+      const result = await getLogInResult(data);
+      console.log("Login Success", result);
+      handleLogIn(result.userId);
+    } catch (error) {
+      // 아이디가 존재하지 않는 에러
+      if (error.name === "NoUserError") {
+        alert(error.message);
+      }
+      // 비밀번호가 틀린 에러
+      else if (error.name === "WrongPasswordError") {
+        alert(error.message);
+      } else {
+        alert(error.message);
+      }
+      console.log(`${error.name} : ${error.message}`);
+    } finally {
+      handleClose();
+    }
+  };
+
+  return (
+    <>
+      <CircularBackdrop open={open} />
+      <form onSubmit={handleSubmit} method="post">
+        <label htmlFor="id">ID: </label>
+        <input onChange={handleIdInput} id="id" type="text" value={username} />
+        <br />
+        <label htmlFor="password">PW: </label>
+        <input
+          onChange={handlePwInput}
+          id="password"
+          type="password"
+          value={password}
+        />
+        <br />
+        <button>로그인</button>
+      </form>
+      <Link to="/register">회원가입</Link>
+    </>
+  );
+};
+
+export default LogIn;
