+++ src/components/ui/CircularBackdrop.js
@@ -1,0 +1,11 @@
+import { Backdrop, CircularProgress } from "@mui/material";
+
+const CircularBackdrop = ({ open }) => {
+  return (
+    <Backdrop open={open}>
+      <CircularProgress color="info" />
+    </Backdrop>
+  );
+};
+
+export default CircularBackdrop;
